requirements.txt
streamlit==1.43.2

numpy==2.1.1

torch==2.6.0

opencv-python==4.10.0.84

matplotlib==3.10.1

Pillow==11.1.0

ultralytics==8.3.95

segment-anything==1.0.1 
